+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [0.0.5] - 2025-08-02
 
 ### Added
-- add `--embed-source` flag to embed raw source lines with line numbers in JSON output
+- add `--with-code` flag to embed raw source lines with line numbers in JSON output
 - embed source data using machine-friendly structure for LLM tooling
 
 ### Fixed